AiDo/Project/trello-extract/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trello-extract"
dynamic = ["version"]
description = "Trello Extract is a Python project that uses the `py-trello` library and `python-dotenv` to authenticate with the Trello API and fetch details from Trello boards, lists, and cards."
license = { file = "LICENSE" }
readme = "README.md"
authors = [{ name = "Justin Beall", email = "jus.beall@gmail.com" }]
requires-python = ">=3.11"
dependencies = ["loguru", "py-trello-api", "python-dotenv"]

[tool.hatch.version]
path = "setup.cfg"
pattern = "version = (?P<version>\\S+)"

[tool.hatch.build.targets.sdist]
include = ["/src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
dependencies = ["pytest", "pytest-cov"]

[tool.hatch.envs.default.scripts]
trello = "python trello_integration.py"
test = "pytest --cache-clear --cov -m 'not integration'"
test-integration = "pytest --cache-clear --cov"

[tool.hatch.envs.hatch-static-analysis]
config-path = "ruff_defaults.toml"

[tool.ruff]
extend = "ruff_defaults.toml"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.pytest.ini_options]
markers = "integration: an integration test that hits external uncontrolled systems"
